backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.3.0
django-filter==2.4.0
psycopg2-binary==2.9.3
python-dotenv